requirements.txt
opencv-python-headless
tensorflow
torch
numpy
pandas
Pillow
